@@ -1,16 +1,48 @@
 {
-  "name": "renderer",
-  "private": true,
-  "version": "0.0.0",
-  "main": "src/main/main.js",
+  "name": "itBuilder",
+  "productName": "itBuilder",
+  "version": "1.0.0",
+  "description": "AI-Powered Automated Development Desktop App",
+  "main": ".vite/build/main.js",
   "scripts": {
-    "dev": "concurrently \"vite\" \"electron-forge start\"",
-    "electron": "electron .",
-    "start": "node scripts/dev.js",
-    "build": "vite build && electron-forge package",
+    "start": "electron-forge start",
     "package": "electron-forge package",
-    "make": "electron-forge make"
+    "make": "electron-forge make",
+    "publish": "electron-forge publish",
+    "lint": "eslint --ext .ts,.tsx ."
   },
+  "devDependencies": {
+    "@electron-forge/cli": "^7.6.1",
+    "@electron-forge/maker-deb": "^7.6.1",
+    "@electron-forge/maker-rpm": "^7.6.1",
+    "@electron-forge/maker-squirrel": "^7.6.1",
+    "@electron-forge/maker-zip": "^7.6.1",
+    "@electron-forge/plugin-auto-unpack-natives": "^7.6.1",
+    "@electron-forge/plugin-fuses": "^7.6.1",
+    "@electron-forge/plugin-vite": "^7.6.1",
+    "@electron/fuses": "^1.8.0",
+    "@types/electron-squirrel-startup": "^1.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "electron": "34.2.0",
+    "eslint": "^8.0.1",
+    "eslint-plugin-import": "^2.25.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3",
+    "vite": "^5.0.12",
+    "vue-tsc": "^2.2.2"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ENATION-UK/itbuilder.git"
+  },
+  "keywords": [],
+  "author": {
+    "name": "kingapex",
+    "email": "kingapex@javashop.cn"
+  },
+  "license": " AGPL-3.0",
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
     "@vicons/fluent": "^0.13.0",
@@ -23,26 +55,5 @@
     "vite-svg-loader": "^5.1.0",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
-  },
-  "devDependencies": {
-    "@electron-forge/cli": "^7.6.1",
-    "@electron-forge/maker-deb": "^7.6.1",
-    "@electron-forge/maker-rpm": "^7.6.1",
-    "@electron-forge/maker-squirrel": "^7.6.1",
-    "@electron-forge/maker-zip": "^7.6.1",
-    "@electron-forge/plugin-auto-unpack-natives": "^7.6.1",
-    "@electron-forge/plugin-fuses": "^7.6.1",
-    "@electron-forge/plugin-vite": "^7.6.1",
-    "@electron/fuses": "^1.8.0",
-    "@types/node": "^22.13.1",
-    "@vitejs/plugin-vue": "^5.2.1",
-    "@vue/tsconfig": "^0.7.0",
-    "concurrently": "^9.1.2",
-    "cross-env": "^7.0.3",
-    "electron": "^34.0.2",
-    "typescript": "~5.6.2",
-    "vite": "^6.0.5",
-    "vue-tsc": "^2.2.0",
-    "wait-on": "^8.0.2"
   }
 }
